Chore: move priority queue to planner package and remove graph pkg
